Fixed bug in tournament matches update

+++ collector/collect_tournament_matches.py
@@ -23,7 +23,12 @@ def print_tournament(tournament, index, count, skip=False):
     )
 
 
-def process_tournament_matches(db, tournament_id):
+def process_tournament_matches(db, tournament, update_only, index, tournaments_count):
+    tournament_id = tournament["_id"]
+    if (update_only and db.aft_matches.count_documents({"tournament id": tournament["_id"]}) > 0):
+        print_tournament(tournament, index, tournaments_count, skip=True)
+        return
+    print_tournament(tournament, index, tournaments_count)
     db.aft_matches.delete_many({"tournament id": tournament_id})
     matches = aft_tournament.get_tournament_matches(tournament_id)
     if matches:
@@ -31,7 +36,12 @@ def process_tournament_matches(db, tournament_id):
         db.aft_matches.insert_many(matches)
 
 
-def process_tournament_draws(db, tournament_id):
+def process_tournament_draws(db, tournament, update_only, index, tournaments_count):
+    tournament_id = tournament["_id"]
+    if (update_only and db.aft_tournament_draws.count_documents({"tournament id": tournament["_id"]}) > 0):
+        print_tournament(tournament, index, tournaments_count, skip=True)
+        return
+    print_tournament(tournament, index, tournaments_count)
     db.aft_tournament_draws.delete_many({"tournament id": tournament_id})
     matches = list(aft_tournament.get_tournament_draws(tournament_id))
     if matches:
@@ -58,16 +68,13 @@ def main(arguments):
     tournaments_count = all_tournaments.count()
     for index, tournament in enumerate(all_tournaments):
         try:
-            if (
-                arguments.update
-                and db.aft_matches.count_documents({"tournament id": tournament["_id"]})
-                > 0
-            ):
-                print_tournament(tournament, index, tournaments_count, skip=True)
-            else:
-                print_tournament(tournament, index, tournaments_count)
-                process_tournament(db, tournament["_id"])
+            process_tournament(db, tournament, arguments.update, index, tournaments_count)
         except Exception as error:
+            print(
+                "{} {} {} ==> {}".format(
+                    tournament["_id"], tournament["name"], tournament["date"], error
+                )
+            )
             logging.error(
                 "{} {} {} ==> {}".format(
                     tournament["_id"], tournament["name"], tournament["date"], error
